cmd/script/requirements.txt
kafka-python
requests
